[feat] PayUser 탈퇴 로직 변경: 로우 삭제 -> withdrawnYn

+++ src/main/kotlin/com/chatandpay/api/service/PayUserService.kt
@@ -25,13 +25,21 @@ class PayUserService(
     @Transactional
     fun register(payUser : SignUpPayUserDTO): PayUserDTO? {
 
-        if(payUserRepository.findByCi(payUser.ci) != null) {
+        val findUser = payUserRepository.findByCi(payUser.ci)
+
+        if(findUser != null && !findUser.isWithdrawn) {
             throw IllegalArgumentException("기 가입 유저입니다.")
         }
 
         val user = userRepository.findById(payUser.userId) ?: throw IllegalArgumentException("존재하지 않는 유저입니다.")
 
-        val regUser = PayUser(ci = payUser.ci, user = user, userSeqNo = payUser.userSeqNo, wallet = null, birthDate = payUser.birthDate)
+        val regUser = if(findUser?.isWithdrawn == true){
+            findUser.withdrawnYn = "N"
+            findUser
+        } else {
+            PayUser(ci = payUser.ci, user = user, userSeqNo = payUser.userSeqNo, wallet = null, birthDate = payUser.birthDate)
+        }
+
         val savedUser = payUserRepository.save(regUser) ?: throw RestApiException("페이 회원 가입에 실패하였습니다.")
         val wallet = Wallet(money = 0, payUser = savedUser)
         savedUser.wallet = walletRepository.save(wallet) ?: throw RestApiException("페이 회원 가입에 실패하였습니다.")
@@ -54,9 +62,16 @@ class PayUserService(
 
         try {
             if (findUserWallet != null) {
-                walletRepository.delete(findUserWallet)
+                val isWalletRemoved = walletRepository.delete(findUserWallet)
+                if(isWalletRemoved) {
+                    findUser.user = null
+                    findUser.wallet = null
+                    findUser.withdrawnYn = "Y"
+                    return payUserRepository.delete(findUser)
+                }
             }
-            return payUserRepository.delete(findUser)
+
+            return false
 
         } catch (e : Exception) {
             throw RestApiException(e.message)
